fix(ci): 修掉 golangci-lint 与 shellcheck 报出的 5 处问题
这两个 job 在 982f303 与 f7196be 上失败。问题都在本轮之前新建、
而由本次提交带上远端的文件里。

## golangci-lint
- internal/gateway/adminapi/admin_patch.go：`"slices"` 被生成脚本分进了非标准库
  分组（我的 import 推导表把 `slices` 漏在 STDLIB 之外）。goimports 要求标准库
  同组 —— 移入第 1 组。
- internal/gateway/count_accounting_test.go:70,96：`int64(got)` 中 got 已是
  int64，属冗余转换（unconvert），去掉。

## shellcheck -S warning（scripts/local-e2e-check.sh）
- SC2164：`cd "$(dirname "$0")/.."` 未处理失败 → 补 `|| exit 1`。
- SC2034：`REPO="$(pwd)"` 全脚本从未被引用 → 删除。
- SC2155：`export VAR="$(...)"` 会掩盖返回值 → 拆成先赋值再 export。
- SC2034 ×2：两处 `for i in $(seq ...)` 的 `i` 都未被使用 → 改用 `_`。

## 验证方式（不靠「本地看着对」）
在本地装出与 CI 相同的三个工具，用**逐字相同的调用**复现门禁：
- `goimports -l internal/ cmd/ test/` → 空；并做变异自证：故意把 `slices`
  再次错放，goimports 立刻报出 admin_patch.go，确认这个「空」不是假绿。
- `unconvert -v ./internal/... ./cmd/... ./test/...` → 无输出。
- `shellcheck -S warning -x scripts/*.sh` → 退出码 0。

另：gofmt 干净；go build / go vet 通过；go test ./... -count=1 全绿。

+++ internal/gateway/count_accounting_test.go
@@ -67,7 +67,7 @@ func TestCount_响应超限时流水仍记实扣量(t *testing.T) {
 	if len(recs) != 1 {
 		t.Fatalf("流水条数 = %d, 期望 1", len(recs))
 	}
-	if got := recs[0].CountUnits; int64(got) != used {
+	if got := recs[0].CountUnits; got != used {
 		t.Errorf("流水 count_units = %d, Redis 实扣 = %d —— 账面与配额不一致（偏差方向=少记）",
 			got, used)
 	}
@@ -93,7 +93,7 @@ func TestCount_成功路径流水与实扣一致(t *testing.T) {
 	if len(recs) != 1 {
 		t.Fatalf("流水条数 = %d, 期望 1", len(recs))
 	}
-	if got := recs[0].CountUnits; int64(got) != used {
+	if got := recs[0].CountUnits; got != used {
 		t.Errorf("流水 count_units = %d, Redis 实扣 = %d", got, used)
 	}
 }
